Gladius Replaced to Reinforced.
Changes it so the Machete Gladius is replaced with the Reinforced machete and the recipe now makes it so the basic machete that is crafted is now reinforced, instead of creating a new one. The metal cost is the same when totaled up but now the original machete is needed instead of the two wood.