Legion Overhaul and FEV Changes
Legion has been changed to a mostly MELEE-BASED system, with very few automatic weapons.

They have been given a speed boost to compensate.

Explorers have been removed and replaced with Primes.

Decanii now choose which branch they want to be apart of: Recruit, Prime, or Veteran, with their own helmet and armor set.

Prime Decanii was given a severely nerfed and de-scoped infiltrator.

Ranger-Hunter Centurion added, with a thermic lance and a huge speed buff.

Most reciped requiring FEV have been changed to require stable mutation toxin, which are findable in the barrels around the map.

Added the molecular destabalizer into RnD, but it requires 2 bottles of FEV, which has to be found in a dungeon. This weapon is awful and should not be used by BoS or good people.

Added an Excavator power armor set to the Vault, and somewhere in dungeons.