NCR MP, HT, Gatling Laser
Added NCR MP, with orders to maintain the peace and ensure the NCR are following the letter of the law.

Added NCR Heavy Trooper, replacing the Heavy Sergeant- so there can't be 3 heavy sargeants.

Replaced the M38 with a weaker variant of LMG that uses 5.56.

Added a Gatling Laser to the Senior Paladin. I think I nerfed it too hard, we'll have to see how it performs.

Added a chance of delimb and extra damage to the Veteran Ranger's AMR.

Made a bunch of changes to dungeons, including the reduction of high-end loot but introduction of things like ambrosia, steel caps, and extra FEV.

Added new sprite for the T-60a helmet, and new Centurion armor for the paladin-slayer Loadout.