NCR Map Update/Minor Fixes
Added a whole new NCR Base, courtesy of Peppers on the Discord.

Fixes the fact that bluespace crystals weren't appearing in the mines >.>

Changed emitters to "Siege Laser" which now destroy walls and do 100 damage.

Nerfed advanced power armor to be actually feasible in combat and put one set somewhere on the map.

Nerfed gatling laser's damage to account for the speed of its fire.