AER14/LAER
This PR's main focus is the LAER and the AER14, as part of my energy weapons update project.

The LAER was difficult to make, and is my own sprite; one I'm quite proud of. It's a bit skinny so I'll see about making it a bit more thick in the future.

It's a unique weapon spawn/a spawn of the Brotherhood of Steel Elder.

The AER14 was added to replace the gatling laser for the Senior Paladin. It's a suped up AER9, basically, with slightly slower firing speed but packs a greater punch for the same ammunition expenditure.

Fixed up the Gauss rifle. It still uses difficult to procure ammunition, but their bullets travel instantly and they do more damage now. There's also recoil.

Fixed the Pulse Pistol using pulse rifle equip sounds.

Edited loot tables to include the new weapons.

Fixed the anti-materiel rifle not having hitscan.