The Big Ol' Followers of the Apocalypse update/ Minor Changes (#405)
* Raider Power Armor, Crafting bug-fixes

This patches add in Raider Power Armor (Findable in tier 4 armor loot pools).

Fixed bugs with the reinforced raider armor crafting recipes

* PR Merge

Combines PR 397 into this.

* Minor adjustments

Swaps out lethal injection drugs in tactical combat fireaid with a combat stimpak.  Adjusts painspike loadout.

* Compilation erorr fix / Tactical First Aid Kit fix

Addresses a compiler error on the newest branch of code.  Switching out NVG medhud with a normal medhud.

* Followers of the Apocolypse Update, other minior changes

This update introduced the followers of the apocolypse, a faction dedicated to the treatment and care of the wasteland at large.

Also tweaks molotovs and xanders