The Little 'Ol Waster Update, Map updates, New Armor (#416)
* The Little 'Ol Waster Update, Map updates

This gives weapons to waster loadouts, themed to their loadout.  Map has been changed to help balance the vault, and to enrich the new dungeon in the sewers.

* Raider armor update

Adds iconoclast to the game.

* Map update

Includes peppers update to FoA

* Iconoclast to loot pool

Added iconoclast to loot pools to be found around.

* Map vertibird fix

Makes it so people cant steal the vertibird and stow it in a trash can.

* The covert complained update

Covert whined about something, so I changed it so he'd whine less.

* The counter argument update

Put the t-51 set back in, but it's now guarded by a deathclaw spawner.

* Not sure

Clearly there was a change.

* Bug fix

I made an oopsie with spawners.

* Final commit for PR419

Adds a conversion crafting recipe to turn salvaged PA into raider PA if desired, given they are stat identical.