Storage containers tweak vol 1 (#480)
* NCR Sprite Bugfix

Fixes the issue with the NCR exile armor not having the correct sprite. Change Log: Corrects the file path to the right one for NCR exile armor (hotfix)

* Bin reduction update

First update to reduce bins from 30 to 6 held items.  This is the first pass-through.

* Baseball bat recipe readded

Not sure why this got deleted, readded

* Storage containers tweak v1

Adds additional logic to restore them to larger sizes.