The "I will not touch things I'm not supposed to" hotfix (#500)
* NCR Sprite Bugfix

Fixes the issue with the NCR exile armor not having the correct sprite. Change Log: Corrects the file path to the right one for NCR exile armor (hotfix)

* Bin reduction update

First update to reduce bins from 30 to 6 held items.  This is the first pass-through.

* Baseball bat recipe readded

Not sure why this got deleted, readded

* Storage containers tweak v1

Adds additional logic to restore them to larger sizes.

* St Jeromes Update

Update

* The unfuck yourself commit

Fixes all the broken recipes where it was required to have a forge.

Adds in crafting for sledgehammers, and throwing knives

Add a forge to the legion camp.

* I will not touch things hotfix

Minor hotfix.