Anti-cheese update / Map bug fixes / NCR Low Power chem dispenser. (#590)
* Live test conclusion, bug fixes

Concludes live test, resealed area, addressed recipe bug and bug with night stalkers

* Dust zone v.1

First interation of the dust zone PVP area.

* Forgot to save

Just a redention of the map.

* Hotfix

Hotfix that addresses space tiles, and making it harder to cheese.

* Map hot fix

Addresses space tiles in the dust bowl area.

* Map update, requested feature

Allows cosmetic clothing to hold a firearm as commonly requested.  Adds onto the dustbowl arena.

* Minor Tweak

Minor tweak to loot in zone

* Minor tweak 1

Minor tweak one

* Tattered red cloak added

Added tattered red cloak for custom loadout purposes, and made it craftable.  This is essentially a crappy version of the death claw cloak that can pass custom loadout approval.

* Map hotfix

Rebalanced the mobs in dust zone.  That's it.

Testing a ~30 percent reduction.

* Hotfix 2

Hotfix 2, space tile

* Map hotfix

Realized didnt have the right tier, spawned ammo for most locations, have a minor boost to the ammo availability/variety down in the dust bowl, given how it takes to do a clear of the area.

* Map hotfix

Addressed issues with ammo spawns.  Didnt use the right tier for some areas. (Shotgun boxes galore)  Added additonal ammo nodes given there wasnt enough put in originally.

* Minor change 3

Misplaced a mob.

* Character slot increase to 30

Character slot increase to 30

* Legendary deathclaw drops deathclaw gauntlet

60 percent chance that if you kill the legendary deathclaw, it will have drop a deathclaw gauntlet.

* bugfix

bugfix

* Power cells description change based on charge

This change makes it so if a cell has 1 percent or less of charge (1 percent due to the tri-beam using a per-shot charge that makes it divide uneveny) makes the cells name shown as a "used cell".

* Shotgun box un-nerfed

Reverses the issues brought upon when the shotgun boxes changed sprites. Causing their capacity to go from 14 to 8

* Icon sprite bug fix

Fixes the invisible shotgun slug box sprite

* Ammo sprite bug fix

Fixes the invisible bug sprite with shotgun slug boxes

* Anti-cheese update, low power dispenser, map-bug fixes

Addresses issues with ghoul chars able to by-pass the legendary ones.  Adds a low power chem dispenser to the NCR camp, anti-cheeses the rest of the dust bowl, and fixes a minor issue in north bunker with the lights being in a random tile in the chem room.