Prefab Map for Mappers (#909)
This is meant to purely be a resource for mappers, either just to inspire a new structure or to pull a prebuilt one directly if time/resources are short and it fits good enough. Can also be used to retain old designs for faction bases in the event we want to revert that specific base and not every map change since it was replaced.