Fluffybirb (#1020)
* Fluffybirb

Adds Vault 134 as a new, temporary dungeon for players to access and plunder. Said Vault also holds relevent event info that players may have missed. Added all three floors to the Dungeon map, and added a construction site to access said dungeon between legion and followers. Also added more cracks around the desert to reflect the damage from the results of the previous event.

* Fluffybirb

Hopefully this fixes the issues with the previous merge of this map!