Changes to the Oasis clinic and Dustbowl
Hello! The Oasis medical clinic now has some new things! A surgical computer along with the discs for lobotomy, pacification, and brainwashing were added. This helps seperate the role of the oasis clinic from that of the followers base, making it more of a "seedy location" to go to for unorthodox surgeries. The blood bags have also been moved to an area closer to the ladder and the patients upstairs they are used on. Other changes include
A chemical storage fridge and bags
A mediDrobe for additional wardrobe options
A box of pill bottles
A dropper
A grinder
Comfortable seating
1 gas mask
And a medical belt.
The floor space of the Oasis clinic has increased by several tiles to accomidate this resulting in the removal of one bonfire and one tier 3 ammo spawn. A tile of diggable block has also been removed to return the outside hallway.
Finally the display cases in dustbowl were removed and replaced with trophy cases as due to kebab and oasis sharing the same zone a lockdown triggered by breaking the cases would close and bolt all doors in the clinic.