Dustbowl Kebab Changes (#1095)
Adds Legendary Raider mob and places quite alot of them in Kebab Town as to make it not just `walk in and get ezloot yes`, thus also making Dustbowl's Kebab Town more difficult. Also adds a few barricades inside of kebab town as to prevent the raiders attacking the muties or ghouls with their wandering.